+++ src/main/java/cn/handyplus/top/hook/PlayerGuildUtil.java
@@ -28,12 +28,12 @@ public class PlayerGuildUtil {
      * @param playerName 玩家名
      * @return 贡献数量
      */
-    public int getPlayerGuildMoney(String playerName) {
+    public long getPlayerGuildMoney(String playerName) {
         if (!PlayerTop.USE_GUILD || StrUtil.isEmpty(playerName)) {
             return 0;
         }
         Integer playerGuildMoney = PlayerGuildApi.getInstance().getPlayerMoney(playerName);
-        return playerGuildMoney != null ? playerGuildMoney : 0;
+        return playerGuildMoney != null ? playerGuildMoney : 0L;
     }
 
     /**
@@ -42,12 +42,12 @@ public class PlayerGuildUtil {
      * @param playerName 玩家名
      * @return 公会战击杀
      */
-    public int getPlayerGuildKill(String playerName) {
+    public long getPlayerGuildKill(String playerName) {
         if (!PlayerTop.USE_GUILD || StrUtil.isEmpty(playerName)) {
             return 0;
         }
         Integer playerGuildKill = PlayerGuildApi.getInstance().getPlayerKill(playerName);
-        return playerGuildKill != null ? playerGuildKill : 0;
+        return playerGuildKill != null ? playerGuildKill : 0L;
     }
 
     /**
@@ -56,12 +56,12 @@ public class PlayerGuildUtil {
      * @param playerName 玩家名
      * @return 公会战死亡
      */
-    public int getPlayerGuildDie(String playerName) {
+    public long getPlayerGuildDie(String playerName) {
         if (!PlayerTop.USE_GUILD || StrUtil.isEmpty(playerName)) {
             return 0;
         }
         Integer playerGuildDie = PlayerGuildApi.getInstance().getPlayerDie(playerName);
-        return playerGuildDie != null ? playerGuildDie : 0;
+        return playerGuildDie != null ? playerGuildDie : 0L;
     }
 
     /**
@@ -70,13 +70,13 @@ public class PlayerGuildUtil {
      * @param playerName 玩家名
      * @return 捐赠贡献数量
      */
-    public int getPlayerDonatedGuildMoney(String playerName) {
+    public long getPlayerDonatedGuildMoney(String playerName) {
         if (!PlayerTop.USE_GUILD || StrUtil.isEmpty(playerName)) {
             return 0;
         }
         try {
             Integer playerGuildMoney = PlayerGuildApi.getInstance().getPlayerGuildMoney(playerName);
-            return playerGuildMoney != null ? playerGuildMoney : 0;
+            return playerGuildMoney != null ? playerGuildMoney : 0L;
         } catch (Throwable ignored) {
         }
         return 0;
